Move ARM code to its own directory
This commit moves the ARM32 code to it's own directory.
