+++ arch/arm32/arm2_div.h
@@ -17,8 +17,8 @@
 #ifndef __SUBTILIS_ARM2_DIV_H__
 #define __SUBTILIS_ARM2_DIV_H__
 
+#include "../../common/ir.h"
 #include "arm_core.h"
-#include "common/ir.h"
 
 void subtilis_arm2_idiv_add(subtilis_ir_section_t *s,
 			    subtilis_arm_section_t *arm_s,
