+++ arch/arm32/arm_core.c
