+++ arch/arm32/arm_core_test.c
