+++ arch/arm32/arm_encode.c
