+++ arch/arm32/arm_encode.h
