+++ arch/arm32/arm_link.c
