+++ arch/arm32/arm_sub_section.c
