+++ arch/arm32/arm_sub_section.h
@@ -17,8 +17,8 @@
 #ifndef __SUBTILIS_ARM_SUB_SECTION_H
 #define __SUBTILIS_ARM_SUB_SECTION_H
 
+#include "../../common/bitset.h"
 #include "arm_core.h"
-#include "common/bitset.h"
 
 struct subtilis_arm_ss_link_t_ {
 	subtilis_bitset_t int_outputs;
