+++ arch/arm32/arm_walker.c
