+++ arch/arm32/fpa_gen.h
@@ -17,8 +17,8 @@
 #ifndef __SUBTILIS_FPA_GEN_H
 #define __SUBTILIS_FPA_GEN_H
 
+#include "../../common/ir.h"
 #include "arm_core.h"
-#include "common/ir.h"
 
 void subtilis_fpa_gen_movr(subtilis_ir_section_t *s, size_t start,
 			   void *user_data, subtilis_error_t *err);
