+++ runarm.c
@@ -17,8 +17,8 @@
 #include <locale.h>
 #include <stdio.h>
 
-#include "arm_disass.h"
-#include "arm_vm.h"
+#include "arch/arm32/arm_disass.h"
+#include "arch/arm32/arm_vm.h"
 #include "common/buffer.h"
 
 const size_t block_size = 16 * 1024;
