+++ unit_tests.c
@@ -16,11 +16,11 @@
 
 #include <locale.h>
 
-#include "arm_core_test.h"
-#include "arm_reg_alloc_test.h"
+#include "arch/arm32/arm_core_test.h"
+#include "arch/arm32/arm_reg_alloc_test.h"
+#include "arch/arm32/fpa_test.h"
 #include "arm_test.h"
 #include "common/bitset_test.h"
-#include "fpa_test.h"
 #include "frontend/ir_test.h"
 #include "frontend/lexer_test.h"
 #include "frontend/parser_test.h"
